102-fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -9,21 +9,21 @@
  */
 int main(void)
 {
-	int i = 1, f0 = 1, f1 = 2, fi;
+	long int i = 1, f0 = 1, f1 = 2, fi;
 
 	for (; i <= 50; i++)
 	{
 		if (i <= 2)
 		{
-			printf("%d, ", i);
+			printf("%ld, ", i);
 			continue;
 		}
 		fi = f0 + f1;
 		f0 = f1;
 		f1 = fi;
-		printf("%d, ", fi);
+		printf("%ld, ", fi);
 		if (i == 50)
-			printf("%d\n", fi);
+			printf("%ld\n", fi);
 	}
 	return (0);
 }
